+++ tests/screenshot/apps/components/DatePicker.js
@@ -1,7 +1,10 @@
-import DatePicker from '../../../../DatePicker';
 import {generateDate} from '@enact/ui-test-utils/utils';
 
+import DatePicker from '../../../../DatePicker';
 
+import {withConfig} from "./utils";
+
+const jan30 = generateDate('2019-01-30');
 const jan31 = generateDate('2019-01-31');
 
 const DatePickerTests = [
@@ -17,6 +20,15 @@ const DatePickerTests = [
 	{
 		locale: 'ar-SA',
 		component: <DatePicker defaultValue={jan31} disabled />
-	}
+	},
+
+	// *************************************************************
+	// focused
+	// *************************************************************
+	...withConfig({focus: true}, [
+		<DatePicker defaultValue={jan30} />,
+		<DatePicker defaultValue={jan30} disabled />,
+		<DatePicker defaultValue={jan30} noLabel />
+	])
 ];
 export default DatePickerTests;